sports_analytics/requirements.txt
# Core Framework
fastapi
uvicorn[standard]
pydantic
pydantic-settings

# LangChain and AI
langchain
langchain-groq
langchain-pinecone
langchain-community
langchain-text-splitters
langchain-core
langchain-nomic

# Vector Database
pinecone-client>=3.0.0,<4.0.0

# Embeddings
nomic

# Data Processing
pandas
numpy

# Environment and Configuration
python-dotenv

# HTTP Client
httpx
requests

# Utilities
python-multipart
aiofiles

# Development and Testing
pytest
pytest-asyncio
black
flake8

# Monitoring and Logging
structlog

# Web Interface
streamlit>=1.28.0 